fix golangview

+++ language/go/golangView/main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"bufio"
 	"bytes"
+	"flag"
 	"fmt"
 	"io/ioutil"
 	"log"
@@ -22,21 +23,33 @@ import (
 	"github.com/jung-kurt/gofpdf"
 )
 
+var (
+	mutex          sync.Mutex
+	modules        ModulesList
+	subModules     Modules
+	filelist       []string
+	ERRORFILE      map[string]string
+	flameGraphList map[string]string
+)
+var (
+	STARTTIME string
+	ENDTIME   string
+)
+
 const PROFILE string = "conf.yaml"
 const JOURNALCTL string = "/bin/journalctl"
-const STARTTIME string = "2023-08-30 07:15:00"
-const ENDTIME string = "2023-08-30 07:25:00"
 const CPUSERVICE string = "qomolo_pidstat"
+const MEMSERVICE string = "qomolo_mem_monitor"
 const ALLSERVICE string = "qomolo_sar_monitor"
 const PDFFILE string = "cpu.pdf"
 const IMAGEPATH string = "png/"
+const CPUINFO string = "cpuinfo"
+const FLAG string = "1"
 
-var (
-	mutex      sync.Mutex
-	modules    ModulesList
-	subModules Modules
-	filelist   []string
-)
+var Sync struct {
+	wg sync.WaitGroup
+	mu sync.Mutex
+}
 
 type ModulesList struct {
 	MODULESLIST []string `yaml:"moduleslist"`
@@ -55,6 +68,13 @@ type StringTicks struct {
 	Labels []string
 }
 
+func checkParam() {
+	if STARTTIME == "" || ENDTIME == "" {
+		fmt.Println("template: ./flag -s '2023-08-31 14:20:00' -e '2023-08-31 14:30:00'")
+		os.Exit(-1)
+	}
+}
+
 func execShell(s string) (string, error) {
 	cmd := exec.Command("/bin/bash", "-c", s)
 	var out bytes.Buffer
@@ -65,7 +85,7 @@ func execShell(s string) (string, error) {
 	if err != nil {
 		return s + " --> error", err
 	}
-	Str := strings.TrimRight(out.String(), "\n") // remote string enter
+	Str := strings.TrimRight(out.String(), "\n") // remove string enter
 	return Str, err
 }
 
@@ -91,6 +111,20 @@ func checkPath(dirPath string) {
 	}
 }
 
+// Ticks 实现 plot.Ticker 接口
+func (s StringTicks) Ticks(min, max float64) []plot.Tick {
+	ticks := []plot.Tick{}
+	for i, v := range s.Values {
+		// 将数值转换为字符串标签
+		tick := plot.Tick{
+			Value: v.X,
+			Label: s.Labels[i],
+		}
+		ticks = append(ticks, tick)
+	}
+	return ticks
+}
+
 func readConf() {
 	_, err := os.Stat(PROFILE)
 	if os.IsNotExist(err) {
@@ -104,6 +138,23 @@ func readConf() {
 	err2 := yaml.Unmarshal(file, &modules)
 	if err2 != nil {
 		log.Fatal(err2)
+	}
+}
+
+func getPID(Pname string, done func()) {
+	defer done()
+	var pid string
+	cmd := "ps aux | grep " + Pname + " | grep -v grep | awk '{print $2}'"
+	ret, _ := execShell(cmd)
+	Plist := strings.Split(ret, "\n")
+	for _, p := range Plist {
+		if ret, _ := execShell("ps -p " + p + "| wc -l"); ret != FLAG {
+			pid = p
+			break
+		}
+	}
+	if pid != "" {
+		flameGraphList[Pname] = pid
 	}
 }
 
@@ -139,7 +190,7 @@ func visitFile(fp string, fi os.DirEntry, err error) error {
 	return nil
 }
 
-func Handle(source []float64, vtype string) (plotter.XYs, plotter.XYs, []string) {
+func Handle(source []float64) (plotter.XYs, plotter.XYs, []string) {
 	data := make(plotter.XYs, len(source))
 	xLabels := make([]string, 0)
 	for i := 0; i < len(source); i++ {
@@ -155,7 +206,7 @@ func Handle(source []float64, vtype string) (plotter.XYs, plotter.XYs, []string)
 	return data, pts, xLabels
 }
 
-func MakePng(source []float64, save_png string, vtype string) {
+func MakePng(source []float64, save_png string) {
 
 	mutex.Lock()         // 申请互斥锁
 	defer mutex.Unlock() // 在函数结束时释放互斥锁
@@ -170,7 +221,7 @@ func MakePng(source []float64, save_png string, vtype string) {
 	p.X.Label.Text = "time"
 	p.Y.Label.Text = "cpu"
 
-	data, pts, xLabels := Handle(source, vtype)
+	data, pts, xLabels := Handle(source)
 	// 创建一个折线图
 	line, err := plotter.NewLine(pts)
 
@@ -189,8 +240,6 @@ func MakePng(source []float64, save_png string, vtype string) {
 	if err := p.Save(20*vg.Inch, 10*vg.Inch, IMAGEPATH+pngName); err != nil {
 		log.Fatal(err)
 	}
-
-	log.Println("save: ", IMAGEPATH+pngName)
 }
 
 func CpuSplitInformation(filename string) Modules {
@@ -210,11 +259,14 @@ func CpuSplitInformation(filename string) Modules {
 		if err != nil {
 			break
 		}
-		// arm
-		// cpu, _ := strconv.ParseFloat(strings.Join(strings.Split(line, " ")[6:7], " "), 64)
-		cpu, _ := strconv.ParseFloat(strings.Join(strings.Split(line, " ")[7:8], " "), 64)
+		//TODO auto switch
+		var cpu float64
+		if ret, _ := execShell("arch"); ret == "x86_64" {
+			cpu, _ = strconv.ParseFloat(strings.Join(strings.Split(line, " ")[7:8], " "), 64)
+		} else {
+			cpu, _ = strconv.ParseFloat(strings.Join(strings.Split(line, " ")[6:7], " "), 64)
+		}
 		time := strings.Join(strings.Split(line, " ")[2:3], " ")
-		// TODO debug  fmt.Println(cpu, time, filename)
 		subModules.Time = append(subModules.Time, time)
 		subModules.Cpu = append(subModules.Cpu, cpu)
 	}
@@ -247,7 +299,7 @@ func fullSplitInformation(filename string) Modules {
 func createPdf() {
 
 	// 创建pdf  页面尺寸为A4
-	pdf := gofpdf.New("P", "mm", "A4", "") //
+	pdf := gofpdf.New("P", "mm", "A5", "")
 
 	// 添加一页新的页面
 	pdf.AddPage()
@@ -260,13 +312,13 @@ func createPdf() {
 	if err != nil {
 		fmt.Printf("error walking the path %v: %v\n", IMAGEPATH, err)
 	}
-	txtOffset := 100
-	imagesOffset := 50
+	txtOffset := 10
+	imagesOffset := 30
 	for i, file := range filelist {
 		// 添加文本 参数依次是宽度、高度、文本内容
-		pdf.Cell(10, float64(10+txtOffset*i), file[len(IMAGEPATH):len(file)-4])
+		pdf.Cell(10, float64(txtOffset*i+10), file[len(IMAGEPATH):len(file)-4])
 		// 插入图片 参数依次是 X、Y 坐标、宽度、高度、图片文件路径、链接
-		pdf.Image(file, 10, float64(20+imagesOffset*i), 80, 40, false, "", 0, "")
+		pdf.Image(file, 10, float64(imagesOffset*i+(len(filelist)*10)), 100, 20, false, "", 0, "")
 	}
 	// save
 	err = pdf.OutputFileAndClose(PDFFILE)
@@ -275,6 +327,12 @@ func createPdf() {
 		return
 	}
 	fmt.Println("save", PDFFILE)
+	clearTmp()
+}
+
+func Init() {
+	checkParam()
+	readConf()
 }
 
 func cpuInfo() {
@@ -283,18 +341,17 @@ func cpuInfo() {
 	cmd := fmt.Sprintf("%s -u %s --since '%s' --until '%s'", JOURNALCTL, CPUSERVICE, STARTTIME, ENDTIME)
 	ret, _ := execShell(cmd)
 	writeFile(ret, FILE)
+	ERRORFILE = make(map[string]string, len(modules.MODULESLIST))
 	for _, v := range modules.MODULESLIST {
 		getinfo := fmt.Sprintf("cat %s | grep %s", FILE, v)
 		v_info, err := execShell(getinfo)
 		subfile := v + ".log"
 		if v_info == "" { // is empty skip
+			ERRORFILE[v] = err.Error()
 			continue
-		}
-		if err != nil {
-			fmt.Println(err.Error())
 		} else {
 			writeFile(v_info, subfile)
-			MakePng(CpuSplitInformation(subfile).Cpu, v, "cpu")
+			MakePng(CpuSplitInformation(subfile).Cpu, v)
 		}
 	}
 	const fullinfofile string = "sar.log"
@@ -303,36 +360,83 @@ func cpuInfo() {
 	writeFile(ret, fullinfofile)
 	cpucmd := fmt.Sprintf("cat %s | grep %s", fullinfofile, "all")
 	cpuinfo, err := execShell(cpucmd)
+
 	if err != nil {
 		fmt.Println(err.Error())
 	} else {
+
 		writeFile(cpuinfo, fullinfofile+"info.log")
-		MakePng(fullSplitInformation(fullinfofile+"info.log").Total, "cpuinfo", "total")
+		MakePng(fullSplitInformation(fullinfofile+"info.log").Total, CPUINFO)
 	}
 }
 
-// Ticks 实现 plot.Ticker 接口
-func (s StringTicks) Ticks(min, max float64) []plot.Tick {
-	ticks := []plot.Tick{}
-	for i, v := range s.Values {
-		// 将数值转换为字符串标签
-		tick := plot.Tick{
-			Value: v.X,
-			Label: s.Labels[i],
-		}
-		ticks = append(ticks, tick)
+func makeFlameGraph() {
+	Sync.wg.Add(len(modules.MODULESLIST))
+	flameGraphList = make(map[string]string, len(modules.MODULESLIST))
+	for _, v := range modules.MODULESLIST {
+		go getPID(v, Sync.wg.Done)
 	}
-	return ticks
+
 }
 
 func clearTmp() {
-	execShell("rm -rv *.log")
-	execShell("rm -rv png/")
+	if os.Getenv("DELETE") == "false" {
+		return
+	} else {
+		execShell("rm -rv *.log")
+		execShell("rm -rv png/")
+		failModules()
+	}
+}
+
+func failModules() {
+	if len(ERRORFILE) == 0 {
+		return
+	} else {
+		fmt.Println("----- fail list -----")
+		for name, error := range ERRORFILE {
+			fmt.Printf("%-20s %s\n", name, error)
+		}
+		ERRORFILE = nil
+	}
+	fmt.Println("end ------ end")
 }
 
 func main() {
-	readConf()
+	flag.StringVar(&STARTTIME, "s", "", "")
+	flag.StringVar(&ENDTIME, "e", "", "")
+	flag.Parse()
+	Init()
 	cpuInfo()
+	// memInfo()
 	createPdf()
-	clearTmp()
+	makeFlameGraph()
+	Sync.wg.Wait()
+	fmt.Println(flameGraphList)
+	for name, pid := range flameGraphList {
+		go execShell("perf_record " + pid + " 1 " + name)
+	}
+	execShell("make_svg")
 }
+
+// func memInfo() {
+// 	const FILE string = "memory.log"
+// 	cmd := fmt.Sprintf("%s -u %s --since '%s' --until '%s'", JOURNALCTL, MEMSERVICE, STARTTIME, ENDTIME)
+// 	ret, _ := execShell(cmd)
+// 	writeFile(ret, FILE)
+// 	ERRORFILE = make(map[string]string, len(modules.MODULESLIST))
+// 	for _, v := range modules.MODULESLIST {
+// 		getinfo := fmt.Sprintf("cat %s | grep %s", FILE, v)
+// 		v_info, err := execShell(getinfo)
+// 		subfile := v + "-mem.log"
+// 		if v_info == "" { // is empty skip
+// 			continue
+// 		}
+// 		if err != nil {
+// 			fmt.Println(err)
+// 			ERRORFILE[v] = err.Error()
+// 		} else {
+// 			writeFile(v_info, subfile)
+// 		}
+// 	}
+// }
